+++ test/src/commonTest/kotlin/com/github/ajalt/colormath/CssParseTest.kt
@@ -8,8 +8,8 @@ import com.github.ajalt.colormath.model.RGBColorSpaces.BT2020
 import com.github.ajalt.colormath.model.RGBColorSpaces.DisplayP3
 import com.github.ajalt.colormath.model.RGBColorSpaces.ROMM_RGB
 import com.github.ajalt.colormath.model.XYZColorSpaces.XYZ50
-import io.kotest.assertions.throwables.shouldNotThrow
 import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.data.Row2
 import io.kotest.data.blocking.forAll
 import io.kotest.data.row
 import io.kotest.matchers.shouldBe
@@ -52,28 +52,23 @@ class CssParseTest {
         row("rgb(1turn,2,3)"),
         row("rgb(1,2,3,4,5)"),
         row("rgb(1,2 3)"),
-        row("rgb(1%,2,3)"),
-        row("rgb(1,2%,3)"),
-        row("rgb(1,2,3%)"),
         row("rgb(1 2,3)"),
         row("rgb(1 2 3, 4)"),
         row("rgb(1,2,3 / 4)"),
         row("rgb(1,2,3 4)"),
         row("hsl(1%,2%,3%)"),
-        row("hsl(1,2%,3)"),
-        row("hsl(1,2,3%)"),
         row("hsl(1degrees,2%,3%)"),
         row("hsl(1ddeg,2%,3%)"),
         row("hsl(1Deg,2%,3%)"),
         row("color(profoto-rgb 0.4835 0.9167 0.2188)")
     ) {
         shouldThrow<IllegalArgumentException> { Color.parse(it) }
-        shouldNotThrow<Exception> { Color.parseOrNull(it) shouldBe null }
+        Color.parseOrNull(it) shouldBe null
     }
 
     @Test
     @JsName("parseCssColor_clamp")
-    fun `parseCssColor clamp`() = forAll(
+    fun `parseCssColor clamp`() = doTest(
         row("rgb(-1,2,3)", RGB.from255(0, 2, 3)),
         row("rgb(256,2,3)", RGB.from255(255, 2, 3)),
         row("rgb(1,256,3)", RGB.from255(1, 255, 3)),
@@ -84,9 +79,7 @@ class CssParseTest {
         row("hsl(1,-2%,3%)", HSL(1, 0, .03)),
         row("hsl(1,2%,-3%)", HSL(1, .02, 0)),
         row("hsl(1,2%,3%,-4%)", HSL(1, .02, .03, 0f)),
-    ) { it, ex ->
-        Color.parse(it) shouldBe ex
-    }
+    )
 
     // Cases mostly from https://developer.mozilla.org/en-US/docs/Web/CSS/color_value
     @Test
@@ -110,12 +103,8 @@ class CssParseTest {
     @Test
     @JsName("parseCssColor_float_exponents")
     fun `parseCssColor float exponents`() {
-        Color.parse("rgb(1e2, .5e1, .5e0, +.25e2%)") shouldBe RGB(
-            100 / 255f,
-            5 / 255f,
-            1 / 255f,
-            .25
-        )
+        Color.parse("rgb(1e2, .5e1, .5e0, +.25e2%)")
+            .shouldBe(RGB(100 / 255f, 5 / 255f, 1 / 255f, .25))
     }
 
     @Test
@@ -137,28 +126,40 @@ class CssParseTest {
         Color.parse(color) shouldBe RGB(51 / 255f, 170 / 255f, 51 / 255f, alpha)
     }
 
+    // https://github.com/web-platform-tests/wpt/blob/master/css/css-color/parsing/color-valid.html
+    @Test
+    @JsName("parseCssColor_rgb")
+    fun `parseCssColor rgb`() = doTest(
+        row("#234", RGB.from255(34, 51, 68)),
+        row("#FEDCBA", RGB.from255(254, 220, 186)),
+        row("rgb(100%, 0%, 0%)", RGB.from255(255, 0, 0)),
+        row("rgba(2, 3, 4, 50%)", RGB.from255(2, 3, 4).copy(alpha = .5f)),
+        row("rgb(-2, 3, 4)", RGB.from255(0, 3, 4)),
+        row("rgb(100, 200, 300)", RGB.from255(100, 200, 255)),
+        row("rgb(20, 10, 0, -10)", RGB.from255(20, 10, 0, 0)),
+        row("rgb(100%, 200%, 300%)", RGB.from255(255, 255, 255)),
+    )
+
     @Test
     @JsName("parseCssColor_hsl")
-    fun `parseCssColor hsl`() = forAll(
-        row("hsl(270,60%,70%)", 270, .6, .7, 1f),
-        row("hsl(270, 60%, 70%)", 270, .6, .7, 1f),
-        row("hsl(270 60% 70%)", 270, .6, .7, 1f),
-        row("hsl(270deg, 60%, 70%)", 270, .6, .7, 1f),
-        row("hsl(4.71239rad, 60%, 70%)", 270, .6, .7, 1f),
-        row("hsl(.75turn, 60%, 70%)", 270, .6, .7, 1f),
-        row("hsl(270, 60%, 50%, .15)", 270, .6, .5, .15),
-        row("hsl(270, 60%, 50%, 15%)", 270, .6, .5, .15),
-        row("hsl(270 60% 50% / .15)", 270, .6, .5, .15),
-        row("hsl(270 60% 50% / 15%)", 270, .6, .5, .15),
-        row("hsla(240, 100%, 50%, .05)", 240, 1.0, .5, .05),
-        row("hsla(240, 100%, 50%, .4)", 240, 1.0, .5, .4),
-        row("hsla(240, 100%, 50%, .7)", 240, 1.0, .5, .7),
-        row("hsla(240, 100%, 50%, 1)", 240, 1.0, .5, 1),
-        row("hsla(240 100% 50% / .05)", 240, 1.0, .5, .05),
-        row("hsla(240 100% 50% / 5%)", 240, 1.0, .5, .05)
-    ) { color, h, s, l, alpha ->
-        Color.parse(color).shouldEqualColor(HSL(h, s, l, alpha))
-    }
+    fun `parseCssColor hsl`() = doTest(
+        row("hsl(270,60%,70%)", HSL(270, .6, .7, 1f)),
+        row("hsl(270, 60%, 70%)", HSL(270, .6, .7, 1f)),
+        row("hsl(270 60% 70%)", HSL(270, .6, .7, 1f)),
+        row("hsl(270deg, 60%, 70%)", HSL(270, .6, .7, 1f)),
+        row("hsl(4.71239rad, 60%, 70%)", HSL(270, .6, .7, 1f)),
+        row("hsl(.75turn, 60%, 70%)", HSL(270, .6, .7, 1f)),
+        row("hsl(270, 60%, 50%, .15)", HSL(270, .6, .5, .15)),
+        row("hsl(270, 60%, 50%, 15%)", HSL(270, .6, .5, .15)),
+        row("hsl(270 60% 50% / .15)", HSL(270, .6, .5, .15)),
+        row("hsl(270 60% 50% / 15%)", HSL(270, .6, .5, .15)),
+        row("hsla(240, 100%, 50%, .05)", HSL(240, 1.0, .5, .05)),
+        row("hsla(240, 100%, 50%, .4)", HSL(240, 1.0, .5, .4)),
+        row("hsla(240, 100%, 50%, .7)", HSL(240, 1.0, .5, .7)),
+        row("hsla(240, 100%, 50%, 1)", HSL(240, 1.0, .5, 1)),
+        row("hsla(240 100% 50% / .05)", HSL(240, 1.0, .5, .05)),
+        row("hsla(240 100% 50% / 5%)", HSL(240, 1.0, .5, .05))
+    )
 
     @Test
     @JsName("parseCssColor_hsl_angles")
@@ -192,82 +193,74 @@ class CssParseTest {
     @Test
     @JsName("parseCssColor_lab")
     // https://www.w3.org/TR/css-color-4/#funcdef-lab
-    fun `parseCssColor lab`() = forAll(
+    fun `parseCssColor lab`() = doTest(
         row("lab(29.2345% 39.3825 20.0664)", LAB50(29.2345, 39.3825, 20.0664)),
         row("lab(52.2345% 40.1645 59.9971)", LAB50(52.2345, 40.1645, 59.9971)),
         row("lab(60.2345% -5.3654 58.956)", LAB50(60.2345, -5.3654, 58.956)),
         row("lab(62.2345% -34.9638 47.7721)", LAB50(62.2345, -34.9638, 47.7721)),
         row("lab(67.5345% -8.6911 -41.6019)", LAB50(67.5345, -8.6911, -41.6019)),
-    ) { str, lab ->
-        Color.parse(str).shouldEqualColor(lab)
-    }
+    )
 
     @Test
     @JsName("parseCssColor_lch")
     // https://www.w3.org/TR/css-color-4/#funcdef-lch
-    fun `parseCssColor lch`() = forAll(
+    fun `parseCssColor lch`() = doTest(
         row("lch(29.2345% 44.2 27)", LCHab50(29.2345, 44.2, 27.0)),
         row("lch(52.2345% 72.2 56.2)", LCHab50(52.2345, 72.2, 56.2)),
         row("lch(60.2345% 59.2 95.2)", LCHab50(60.2345, 59.2, 95.2)),
         row("lch(62.2345% 59.2 126.2)", LCHab50(62.2345, 59.2, 126.2)),
         row("lch(67.5345% 42.5 258.2)", LCHab50(67.5345, 42.5, 258.2)),
-    ) { str, lch ->
-        Color.parse(str).shouldEqualColor(lch)
-    }
+    )
 
     @Test
     @JsName("parseCssColor_hwb")
     // No examples in the spec for this one
-    fun `parseCssColor hwb`() = forAll(
+    fun `parseCssColor hwb`() = doTest(
         row("hwb(180 0% 0%)", HWB(180.0, 0.0, 0.0)),
         row("hwb(180deg 23.4% 45.6%)", HWB(180.0, .234, .456)),
         row("hwb(200grad 23.4% 45.6%)", HWB(180.0, .234, .456)),
         row("hwb(0.5turn 23.4% 45.6%)", HWB(180.0, .234, .456)),
         row("hwb(3.1416rad 23.4% 45.6%)", HWB(180.0, .234, .456)),
-    ) { str, hwb ->
-        Color.parse(str).shouldEqualColor(hwb)
-    }
+    )
 
     @Test
     @JsName("parseCssColor_oklab")
     // https://www.w3.org/TR/css-color-4/#ex-oklab-samples
-    fun `parseCssColor oklab`() = forAll(
+    fun `parseCssColor oklab`() = doTest(
         row("oklab(40.101%  0.1147  0.0453)", Oklab(0.40101, 0.1147, 0.0453)),
         row("oklab(59.686%  0.1009  0.1192)", Oklab(0.59686, 0.1009, 0.1192)),
         row("oklab(0.65125 -0.0320  0.1274)", Oklab(0.65125, -0.0320, 0.1274)),
         row("oklab(66.016% -0.1084  0.1114)", Oklab(0.66016, -0.1084, 0.1114)),
         row("oklab(72.322% -0.0465 -0.1150)", Oklab(0.72322, -0.0465, -0.1150)),
-//        row("oklab(42.1% 41% -25%)", Oklab()), TODO: a/b percentage scaling
-    ) { str, lab ->
-        Color.parse(str).shouldEqualColor(lab)
-    }
+        row("oklab(51.975% -35.075% 26.92%)", RGB("#008000").toOklab()),
+    )
 
     @Test
     @JsName("parseCssColor_oklch")
     // https://www.w3.org/TR/css-color-4/#ex-oklch-samples
-    fun `parseCssColor oklch`() = forAll(
+    fun `parseCssColor oklch`() = doTest(
         row("oklch(40.101% 0.12332 21.555)", Oklch(0.40101, 0.12332, 21.555)),
         row("oklch(59.686% 0.15619 49.7694)", Oklch(0.59686, 0.15619, 49.7694)),
         row("oklch(0.65125 0.13138 104.097)", Oklch(0.65125, 0.13138, 104.097)),
         row("oklch(0.66016 0.15546 134.231)", Oklch(0.66016, 0.15546, 134.231)),
         row("oklch(72.322% 0.12403 247.996)", Oklch(0.72322, 0.12403, 247.996)),
         row("oklch(0% 0 none)", Oklch(0, 0, NaN)),
-//        row("oklch(42.1% 48.25% 328.4)", Oklch()), TODO: c percentage scaling
-    ) { str, lch ->
-        Color.parse(str).shouldEqualColor(lch)
-    }
+        row("oklch(51.975% 44.215% 142.495)", RGB("#008000").toOklch()),
+    )
 
     @Test
     @JsName("parseCssColor_color")
     // https://www.w3.org/TR/css-color-4/#funcdef-color
-    fun `parseCssColor color`() = forAll(
+    fun `parseCssColor color`() = doTest(
         row("color(srgb 25% 50% 75% / 90%)", SRGB(.25, .5, .75, .9)),
-        row("color(display-p3 -0.6112 1.0079 -0.2192)", DisplayP3(0f, 1f, 0f)),
+        row("color(display-p3 0 1 0)", DisplayP3(0f, 1f, 0f)),
         row("color(a98-rgb 25% 50% 75%)", AdobeRGB(.25, .5, .75)),
         row("color(prophoto-rgb 25% 50% 75% / 90%)", ROMM_RGB(.25, .5, .75, .9)),
         row("color(rec2020 0.42053 0.979780 0.00579)", BT2020(0.42053, 0.979780, 0.00579)),
         row("color(xyz 0.2005 0.14089 0.4472)", XYZ50(0.2005, 0.14089, 0.4472)),
-    ) { str, lch ->
-        Color.parse(str).shouldEqualColor(lch)
+    )
+
+    private fun doTest(vararg rows: Row2<String, Color>) = forAll(*rows) { str, color ->
+        Color.parse(str).shouldEqualColor(color)
     }
 }